Fixes #9, renaming scripts mistakes
